+++ tests/test_header.py
@@ -94,7 +94,7 @@ def test_metadata_and_cell_to_header2():
 def test_notebook_from_plain_script_has_metadata_filter(script="""print('Hello world")
 """):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):
-        nb = jupytext.reads(script, '.py')
+        nb = jupytext.reads(script, '.py', additional_metadata_on_text_files=False)
     assert nb.metadata.get('jupytext', {}).get('metadata_filter', {}).get('notebook') is False
     assert nb.metadata.get('jupytext', {}).get('metadata_filter', {}).get('cells') is False
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):